feat: Move participant to a new file
