+++ CardGameApp/CardGameApp/Models/Players.swift
@@ -34,22 +34,3 @@ class Players {
         return players.map { transform($0) }
     }
 }
-
-class Participant {
-    private var cards = [Card]()
-    
-    func take(card: Card) {
-        cards.append(card)
-    }
-    
-    @discardableResult
-    func repeatForEachCard<T>(_ transform: (Card) -> T) -> [T] {
-        return cards.map { transform($0) }
-    }
-}
-
-extension Participant: Equatable {
-    static func == (lhs: Participant, rhs: Participant) -> Bool {
-        return lhs.cards == rhs.cards
-    }
-}
